backend/requirements.txt
transformers
torchaudio
torch
sentencepiece
protobuf
accelerate
pyannote-audio
faster-whisper
alembic
pydantic
pydantic-settings
python-dotenv
uvicorn
gunicorn
python-multipart
fastapi
sqlalchemy
asyncpg
uvloop
websockets
sentry-sdk[fastapi]
langfuse
pydub